modify to fix many modal btn on same page
