+++ app/javascript/scripts/index.js
@@ -1,1 +1,1 @@
-import './ShareModal'
+import './modal'
