+++ app/javascript/scripts/modal.js
@@ -1,0 +1,54 @@
+//於頁面點擊.modal-btn時，出現蓋板遮罩與視窗，視窗會撈取本頁面的template
+document.addEventListener("DOMContentLoaded", function() {
+  const body = document.querySelector('body')
+  const modalBtn = document.querySelectorAll('.modal-btn')
+  var modalClosing
+
+  if(modalBtn){
+    modalBtn.forEach( btn => {
+    btn.addEventListener('click', (e) => {
+      const modalTarget = e.target
+      const btnTemplate = getBtnTemplate(modalTarget, 1) //固定寫在class list裡面第二個
+      modalTarget.setAttribute('disabled', 'true')
+
+      let scroll = window.scrollY //存下現在的scrollbar Y值
+      document.body.style.position = 'fixed'
+      document.body.style.top = `-${scroll}px`
+
+      const modalDiv = document.createElement('div')
+      modalDiv.innerHTML = document.querySelector(`.${btnTemplate}`).innerHTML
+
+      const mask = modalDiv.querySelector('.modal-fade')
+      const modal = modalDiv.querySelector('.modal-content')
+
+      body.appendChild(modalDiv)
+      setTimeout(() => {modal.classList.toggle('modal-show')}, 300);
+      setTimeout(()=>{mask.classList.toggle('mask-show')},10)
+
+      mask.addEventListener('click', () => {
+        if (!modalClosing) {
+          modalClosing = true
+          console.log('close')
+          setTimeout(() => {
+            modal.classList.toggle('modal-show')
+            mask.classList.toggle('mask-show')            
+          }, 300);
+          setTimeout(()=>{ 
+            body.removeChild(modalDiv)
+            modalClosing = false 
+          },1000)
+          document.body.style.position = ''
+          document.body.style.top = ''
+          window.scrollTo(0, scroll || 0)
+          btn.removeAttribute('disabled')
+        } else {
+          return
+        }
+      })
+    })
+  })
+  }
+  function getBtnTemplate(node, n) {
+    return node.classList[n].substring(0, node.classList[n].length-4) + '-template'
+  }
+})
