+++ app/javascript/styles/index.js
@@ -1,1 +1,1 @@
-import './share_modal.css'
+import './modal.css'
