+++ app/views/songs/show.html.erb
@@ -1,5 +1,6 @@
 <h1><%=@song.name%></h1>
-<button id="share-song">Share</button><br>
+<button class="modal-btn share-btn">Share</button><br>
+<button class="modal-btn addlist-btn">Add to List</button><br>
 
 
   <ul>
@@ -37,3 +38,29 @@
   <% end %>
 
   <div id="player1" class="aplayer"></div>
+
+<% #以下之後用Vue藏內容應該可以 %>
+<template class="share-template">
+  <div class="modal-container share-container">
+    <div class="modal-fade">
+      <div class="modal-mask">
+    </div>
+    <button class="modal-close-btn"> X </button>
+  </div>
+  <div class="modal-content">1
+  </div>
+  </div>
+</template>
+
+<template class="addlist-template">
+  <div class="modal-container addlist-container">
+    <div class="modal-fade">
+      <div class="modal-mask">
+    </div>
+    <button class="modal-close-btn"> X </button>
+  </div>
+  <div class="modal-content">1
+  </div>
+  </div>
+</template>
+>>>>>>> modify to fix many modal btn on same page
